Move shoot-specific utils into `pkg/utils/gardener` package
